+++ packages/components-vue/src/examples/InputMonthExampleControlled.vue
@@ -1,16 +1,16 @@
 <script setup lang="ts">
-  import { type InputMonthInputEventDetail, PInputMonth, PText } from '@porsche-design-system/components-vue';
-  import { computed, ref } from 'vue';
+import { type InputMonthInputEventDetail, PInputMonth, PText } from '@porsche-design-system/components-vue';
+import { computed, ref } from 'vue';
 
-  const value = ref<string>('');
+const value = ref<string>('');
 
-  const onInput = (e: InputMonthInputEventDetail) => {
-    value.value = (e.target as HTMLInputElement).value;
-  };
+const onInput = (e: CustomEvent<InputMonthInputEventDetail>) => {
+  value.value = (e.target as HTMLInputElement).value;
+};
 
-  const debugText = computed(() => {
-    return `Value: ${value.value}`;
-  });
+const debugText = computed(() => {
+  return `Value: ${value.value}`;
+});
 </script>
 
 <template>