+++ packages/components/src/components/input-month/input-month.tsx
@@ -125,6 +125,9 @@ export class InputMonth {
 
   @Watch('value')
   public onValueChange(newValue: string): void {
+    if (this.inputElement && this.inputElement.value !== newValue) {
+      this.inputElement.value = newValue;
+    }
     this.internals?.setFormValue(newValue);
   }
 
@@ -219,6 +222,7 @@ export class InputMonth {
         {...(hasShowPickerSupport() && {
           end: (
             <PrefixedTagNames.pButtonPure
+              tabIndex={this.disabled ? -1 : null}
               hideLabel={true}
               class="button"
               type="button"