+++ packages/components/src/components/input-time/input-time.tsx
@@ -125,6 +125,9 @@ export class InputTime {
 
   @Watch('value')
   public onValueChange(newValue: string): void {
+    if (this.inputElement && this.inputElement.value !== newValue) {
+      this.inputElement.value = newValue;
+    }
     this.internals?.setFormValue(newValue);
   }
 
@@ -219,6 +222,7 @@ export class InputTime {
         {...(hasShowPickerSupport() && {
           end: (
             <PrefixedTagNames.pButtonPure
+              tabIndex={this.disabled ? -1 : null}
               hideLabel={true}
               class="button"
               type="button"